git-svn-id: http://192.168.20.11/sequoiadb/trunk@17525 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/client/clientcpp.cpp
@@ -6733,7 +6733,7 @@ do                                                            \
 
       replyHeader = ( const MsgOpReply * )_pReceiveBuffer ;
       if ( 1 == replyHeader->numReturned &&
-           sizeof( MsgOpReply ) < replyHeader->header.messageLength )
+           (INT32)(sizeof( MsgOpReply )) < replyHeader->header.messageLength )
       {
          try
          {
